Update a Facility in the Database

Given a facility name, a list of fields, and a corresponding list of values, update facility information in the database.

Usage

update_facility(facility_name = NULL, field, value, test_db = FALSE)

Arguments

facility_name

The name of the facility. Must be the primary name.

field

A list of field names whose values you wish to change. Should align with value.

value

A list of values you wish to change. Should align with field.

test_db

Is this a connection to the test database? Defaults to FALSE.

Value

A tibble with one row containing the record of the updated facility.

Examples

  update_facility(
  facility_name = "Buckland Community Centre",
  field = list("physical_address", "designation"),
  value = list("Corner Logan and Buckville Road", "Community Centre"),
  test_db = TRUE
  )
#> # A tibble: 1 × 13
#>   facilit…¹ facil…² facil…³ name  physi…⁴ local…⁵ desig…⁶ deliv…⁷ facil…⁸ closed
#>   <chr>     <chr>   <chr>   <chr> <chr>   <chr>   <chr>   <chr>   <chr>   <lgl> 
#> 1 Asset     A26     FA174   Buck… Cnr Lo… Frankl… Rural … Commun… Privat… FALSE 
#> # … with 3 more variables: staffed <lgl>, leased <lgl>,
#> #   entry_access_type <chr>, and abbreviated variable names ¹​facility_type,
#> #   ²​facility_id, ³​facility_attribute_id, ⁴​physical_address, ⁵​local_board,
#> #   ⁶​designation, ⁷​delivery_model, ⁸​facility_ownership